+++ components/PianoStarGame/PianoStarGame.module.scss
@@ -1,60 +1,61 @@
-@use '../../styles/variables' as *;
-
 .gameSection {
-  margin: 3rem auto;
-  max-width: 40rem;
-  padding: 1.5rem;
-  background-color: #fafafa;
-  border-radius: 1rem;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  padding: 2rem;
   text-align: center;
 }
 
 .title {
   font-size: 1.5rem;
-  font-weight: 600;
   margin-bottom: 1rem;
 }
 
+#sheet {
+  margin: 1rem auto;
+  width: 600px;
+  height: 160px;
+}
+
 .piano {
-  margin-top: 1.5rem;
   display: flex;
   justify-content: center;
-  user-select: none;
   position: relative;
+  height: 160px;
 }
 
 .key {
-  width: 48px;
-  height: 180px;
-  border: 1px solid #999;
-  display: inline-flex;
-  align-items: flex-end;
-  justify-content: center;
-  background: #fafafa;
-  margin: 0 1px;
-  cursor: pointer;
+  background: white;
+  border: 1px solid #ccc;
+  width: 40px;
+  height: 100%;
   position: relative;
-  transition: background 0.05s;
+  text-align: center;
+  font-weight: bold;
+  z-index: 1;
+  user-select: none;
+  display: flex;
+  justify-content: center;
+  align-items: flex-end;
+  padding-bottom: 10px;
+  margin: 0 1px;
 }
 
 .black {
-  width: 36px;
-  height: 120px;
-  background: #111;
-  color: #fff;
+  background: black;
+  color: white;
+  width: 25px;
+  height: 60%;
   position: absolute;
-  left: 32px;
   top: 0;
-  z-index: 1;
+  left: 28px; // shift it over the right half
+  z-index: 2;
 }
 
+
 .pressed {
-  background: $yellow-color !important;
+  background-color: #60a5fa !important;
+  color: white;
 }
 
 .hint {
   margin-top: 1rem;
-  font-size: 0.875rem;
-  color: #666;
+  font-size: 1rem;
 }
