+++ lib/piano-game/constants.ts
@@ -1,0 +1,24 @@
+export const WHITE_KEYS = [
+    { key: 'A', note: 'C4', position: 0 },
+    { key: 'S', note: 'D4', position: 1 },
+    { key: 'D', note: 'E4', position: 2 },
+    { key: 'F', note: 'F4', position: 3 },
+    { key: 'G', note: 'G4', position: 4 },
+    { key: 'H', note: 'A4', position: 5 },
+    { key: 'J', note: 'B4', position: 6 },
+];
+
+export const BLACK_KEYS = [
+    { key: 'W', note: 'C#4', position: 0 },
+    { key: 'E', note: 'D#4', position: 1 },
+    { key: 'T', note: 'F#4', position: 3 },
+    { key: 'Y', note: 'G#4', position: 4 },
+    { key: 'U', note: 'A#4', position: 5 },
+];
+
+export const SONG = [
+    'C4', 'C4', 'G4', 'G4', 'A4', 'A4', 'G4',
+    'F4', 'F4', 'E4', 'E4', 'D4', 'D4', 'C4'
+];
+
+export const IDLE_TIMEOUT_MS = 4500;
