+++ lib/piano-game/types.ts
@@ -1,0 +1,5 @@
+export interface NoteKey {
+    key: string;
+    note: string;
+    position: number;
+}
