+++ package.json
@@ -57,9 +57,10 @@
     "react-hook-form": "latest",
     "react-resizable-panels": "^2.1.7",
     "recharts": "2.15.0",
-    "sass": "^1.86.0",
     "sonner": "^1.7.1",
+    "tone": "^15.1.22",
     "vaul": "^0.9.6",
+    "vexflow": "^5.0.0",
     "zod": "latest"
   },
   "devDependencies": {
@@ -67,6 +68,7 @@
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "postcss": "^8",
+    "sass": "^1.89.2",
     "tailwindcss": "^3.4.17",
     "typescript": "^5"
   }
